Move failover exceptions to dwave.system.exceptions
