+++ dwave/system/exceptions.py
@@ -15,3 +15,15 @@
 
 class MissingFluxBias(Exception):
     """"""
+
+
+class FailoverCondition(Exception):
+    """QPU/SolverAPI call failed with an error that might be mitigated by
+    retrying on a different solver.
+    """
+
+
+class RetryCondition(FailoverCondition):
+    """QPU/SolverAPI call failed with an error that might be mitigated by
+    retrying on the same solver.
+    """
